EFM32 series 2: Add support for debug lock/erase
Add support for interacting with DCI over SWD, making it possible to
issue a device erase if the device is locked (and device erase has not
been disabled), and also to activate the debug lock.

Also attempt to find out if the debug lock is active if target examine
fails, and make some helpful prints.